silent-alarm/requirements.txt
opencv-python
google-generativeai
numpy
python-dotenv
